backend/requirements.txt
bidict==0.23.1
click==8.1.8
colorama==0.4.6
h11==0.14.0
nest-asyncio==1.6.0
python-engineio==4.11.2
python-socketio==5.12.1
simple-websocket==1.1.0
uvicorn==0.34.0
wsproto==1.2.0
